Add a UUID to the NodeRef
This allows the Petri Net to support multiple places or transitions with the same name in a simple manner.

The label must stay in the PlaceRef and TransitionRef so that we can order the elements in the PetriNet alphabetically. Otherwise, if the reference only contained a UUID, the order would be random, similar to using a HashMap instead of a BTreeMap.