Fix LoLA format: List ALL places in the `MARKING` section, even if there are no tokens in those places.
This is important to be consistent with the formatting.
The `MARKING` section should have a semicolon at the end and an empty line, just like the `PLACES` and `TRANSITIONS` sections. The previous version failed to comply with this. By being explicit and listing all places always, we keep the code simple and the formatting correct.